refactor: 리캡 생성 ui/ux 수정
