+++ src/utils/index.ts
@@ -28,3 +28,17 @@ export function fullDateStr() {
   const today = new Date()
   return `${today.getFullYear()}${today.getMonth() + 1}${today.getDate()}`
 }
+
+export const base64ToBlob = (base64Data: string, contentType: string): Blob => {
+  const byteCharacters = atob(base64Data)
+  const byteNumbers = new Array(byteCharacters.length)
+  for (let i = 0; i < byteCharacters.length; i++) {
+    byteNumbers[i] = byteCharacters.charCodeAt(i)
+  }
+  const byteArray = new Uint8Array(byteNumbers)
+  return new Blob([byteArray], { type: contentType })
+}
+
+export const blobToFile = (blob: Blob, fileName: string): File => {
+  return new File([blob], fileName, { type: blob.type })
+}
